backend/requirements.txt
fastapi[all]
uvicorn
tinydb
pydantic
pydantic[email]
sqlalchemy
psycopg2-binary
asyncpg
alembic
pydantic-settings
greenlet
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0
faker>=25.0.0
aiosqlite>=0.19.0
